requirements.txt
httpx==0.28.1
python-telegram-bot==22.0
python-dotenv==1.1.0
pydantic==2.11.3
requests==2.32.3
beautifulsoup4==4.13.4
setuptools==79.0.1
python-dotenv
python-telegram-bot